@@ -5,177 +5,151 @@
         "ng": "ng",
         "start": "node --max_old_space_size=8048 ./node_modules/@angular/cli/bin/ng serve --configuration development --host 0.0.0.0 --open",
         "build": "ng build",
-        "build:prod": "node --max_old_space_size=4096 ./node_modules/@angular/cli/bin/ng build --configuration production --vendor-chunk",
+        "build:prod": "node --max_old_space_size=4096 ./node_modules/@angular/cli/bin/ng build --configuration production",
         "build:types": "node generate-types.js",
-        "test": "ng test",
+        "build:icon-metadata": "node generate-icon-metadata.js",
         "lint": "ng lint",
-        "e2e": "ng e2e",
         "prepare": "patch-package",
         "bundle-report": "ng build --configuration production --source-map && source-map-explorer target/generated-resources/public/*.js"
     },
     "private": true,
     "dependencies": {
-        "@angular/animations": "^15.2.10",
-        "@angular/cdk": "^15.2.9",
-        "@angular/common": "^15.2.10",
-        "@angular/compiler": "^15.2.10",
-        "@angular/core": "^15.2.10",
+        "@angular/animations": "18.2.10",
+        "@angular/cdk": "18.2.11",
+        "@angular/common": "18.2.10",
+        "@angular/compiler": "18.2.10",
+        "@angular/core": "18.2.10",
         "@angular/flex-layout": "^15.0.0-beta.42",
-        "@angular/forms": "^15.2.10",
-        "@angular/material": "^15.2.9",
-        "@angular/platform-browser": "^15.2.10",
-        "@angular/platform-browser-dynamic": "^15.2.10",
-        "@angular/router": "^15.2.10",
-        "@auth0/angular-jwt": "^5.1.2",
-        "@date-io/core": "1.3.7",
-        "@date-io/date-fns": "1.3.7",
+        "@angular/forms": "18.2.10",
+        "@angular/material": "18.2.11",
+        "@angular/platform-browser": "18.2.10",
+        "@angular/platform-browser-dynamic": "18.2.10",
+        "@angular/router": "18.2.10",
+        "@auth0/angular-jwt": "^5.2.0",
         "@flowjs/flow.js": "^2.14.1",
-        "@flowjs/ngx-flow": "~0.6.0",
-        "@geoman-io/leaflet-geoman-free": "2.14.2",
-        "@iplab/ngx-color-picker": "^15.0.2",
-        "@juggle/resize-observer": "^3.4.0",
-        "@mat-datetimepicker/core": "~11.0.3",
-        "@material-ui/core": "4.12.3",
-        "@material-ui/icons": "4.11.2",
-        "@material-ui/pickers": "3.3.10",
-        "@mdi/svg": "^7.2.96",
-        "@messageformat/core": "^3.1.0",
-        "@ngrx/effects": "^15.4.0",
-        "@ngrx/store": "^15.4.0",
-        "@ngrx/store-devtools": "^15.4.0",
-        "@ngx-translate/core": "^14.0.0",
-        "@ngx-translate/http-loader": "^7.0.0",
-        "@tinymce/tinymce-angular": "^7.0.0",
-        "@types/chart.js": "2.9.38",
-        "ace-builds": "1.4.13",
+        "@flowjs/ngx-flow": "18.0.1",
+        "@geoman-io/leaflet-geoman-free": "2.17.0",
+        "@iplab/ngx-color-picker": "^18.0.1",
+        "@mat-datetimepicker/core": "~14.0.0",
+        "@mdi/svg": "^7.4.47",
+        "@messageformat/core": "^3.4.0",
+        "@ngrx/effects": "^18.1.1",
+        "@ngrx/store": "^18.1.1",
+        "@ngrx/store-devtools": "^18.1.1",
+        "@ngx-translate/core": "^15.0.0",
+        "@tinymce/tinymce-angular": "^8.0.1",
+        "ace-builds": "1.36.2",
         "ace-diff": "^3.0.3",
-        "angular-gridster2": "~15.0.4",
-        "angular2-hotkeys": "^13.1.0",
+        "angular-gridster2": "~18.0.1",
+        "angular2-hotkeys": "^16.0.1",
         "canvas-gauges": "^2.1.7",
-        "chart.js": "4.4.0",
-        "chartjs-adapter-moment": "1.0.1",
-        "chartjs-plugin-annotation": "3.0.1",
-        "chartjs-plugin-zoom": "2.0.1",
-        "core-js": "^3.29.1",
-        "date-fns": "2.0.0-alpha.27",
-        "dayjs": "1.11.4",
+        "core-js": "^3.38.1",
+        "dayjs": "1.11.13",
         "flot": "https://github.com/thingsboard/flot.git#0.9-work",
         "flot.curvedlines": "https://github.com/MichaelZinsmaier/CurvedLines.git#master",
         "font-awesome": "^4.7.0",
         "html2canvas": "^1.4.1",
-        "jquery": "^3.6.3",
-        "jquery.terminal": "^2.35.3",
-        "js-beautify": "1.14.7",
+        "jquery": "^3.7.1",
+        "jquery.terminal": "^2.43.1",
+        "js-beautify": "1.15.1",
         "json-schema-defaults": "^0.4.0",
-        "jstree": "^3.3.15",
+        "jstree": "^3.3.17",
         "jstree-bootstrap-theme": "^1.0.1",
         "jszip": "^3.10.1",
-        "leaflet": "1.8.0",
+        "leaflet": "1.9.4",
         "leaflet-polylinedecorator": "1.6.0",
-        "leaflet-providers": "1.13.0",
+        "leaflet-providers": "2.0.0",
         "leaflet.gridlayer.googlemutant": "0.14.1",
         "leaflet.markercluster": "1.5.3",
-        "libphonenumber-js": "^1.10.4",
-        "marked": "^4.0.17",
-        "moment": "^2.29.4",
-        "moment-timezone": "^0.5.42",
-        "mqtt": "~5.3.0",
-        "ngx-clipboard": "^15.1.0",
+        "libphonenumber-js": "^1.11.10",
+        "marked": "~12.0.2",
+        "moment": "^2.30.1",
+        "moment-timezone": "^0.5.45",
+        "ngx-clipboard": "^16.0.0",
         "ngx-daterangepicker-material": "^6.0.4",
-        "ngx-drag-drop": "^15.0.1",
-        "ngx-flowchart": "https://github.com/thingsboard/ngx-flowchart.git#release/2.0.0",
-        "ngx-hm-carousel": "^3.0.0",
-        "ngx-markdown": "^15.1.2",
-        "ngx-sharebuttons": "^12.0.0",
-        "ngx-translate-messageformat-compiler": "^6.2.0",
+        "ngx-drag-drop": "^18.0.2",
+        "ngx-flowchart": "https://github.com/thingsboard/ngx-flowchart.git#release/3.0.0",
+        "ngx-hm-carousel": "^18.0.0",
+        "ngx-markdown": "^18.1.0",
+        "ngx-sharebuttons": "^15.0.3",
+        "ngx-translate-messageformat-compiler": "^7.0.0",
         "objectpath": "^2.0.0",
         "prettier": "^2.8.3",
         "prop-types": "^15.8.1",
-        "qrcode": "^1.5.1",
+        "qrcode": "^1.5.4",
         "raphael": "^2.3.0",
-        "rc-select": "13.2.1",
-        "react": "17.0.2",
-        "react-ace": "9.5.0",
-        "react-dom": "17.0.2",
-        "react-dropzone": "^11.4.2",
+        "rc-select": "14.15.2",
+        "react": "18.3.1",
+        "react-ace": "12.0.0",
+        "react-dom": "18.3.1",
+        "react-dropzone": "14.2.9",
         "reactcss": "^1.2.3",
-        "rxjs": "~7.8.0",
-        "schema-inspector": "^2.0.2",
+        "rxjs": "~7.8.1",
+        "schema-inspector": "^2.1.0",
         "screenfull": "^6.0.2",
         "split.js": "^1.6.5",
-        "systemjs": "6.14.1",
+        "systemjs": "6.15.1",
         "tinycolor2": "^1.6.0",
-        "tinymce": "~5.10.7",
+        "tinymce": "~6.8.5",
         "tooltipster": "^4.2.8",
-        "ts-transformer-keys": "^0.4.4",
-        "tslib": "^2.5.0",
+        "tslib": "^2.7.0",
         "tv4": "^1.3.0",
         "typeface-roboto": "^1.1.13",
-        "zone.js": "~0.13.0"
+        "zone.js": "~0.14.10"
     },
     "devDependencies": {
-        "@angular-builders/custom-webpack": "~15.0.0",
-        "@angular-devkit/build-angular": "^15.2.10",
-        "@angular-eslint/builder": "15.2.1",
-        "@angular-eslint/eslint-plugin": "15.2.1",
-        "@angular-eslint/eslint-plugin-template": "15.2.1",
-        "@angular-eslint/schematics": "15.2.1",
-        "@angular-eslint/template-parser": "15.2.1",
-        "@angular/cli": "^15.2.10",
-        "@angular/compiler-cli": "^15.2.10",
-        "@angular/language-service": "^15.2.10",
-        "@ngtools/webpack": "15.2.1",
-        "@types/ace-diff": "^2.1.1",
-        "@types/canvas-gauges": "^2.1.4",
-        "@types/flot": "^0.0.32",
-        "@types/flowjs": "^2.13.9",
-        "@types/jasmine": "~3.10.2",
-        "@types/jasminewd2": "^2.0.10",
-        "@types/jquery": "^3.5.16",
-        "@types/js-beautify": "^1.13.3",
-        "@types/leaflet": "1.8.0",
+        "@angular-builders/custom-esbuild": "18.0.0",
+        "@angular-devkit/build-angular": "18.2.11",
+        "@angular-eslint/builder": "18.4.0",
+        "@angular-eslint/eslint-plugin": "18.4.0",
+        "@angular-eslint/eslint-plugin-template": "18.4.0",
+        "@angular-eslint/schematics": "18.4.0",
+        "@angular-eslint/template-parser": "18.4.0",
+        "@angular/cli": "18.2.11",
+        "@angular/compiler-cli": "18.2.10",
+        "@angular/language-service": "18.2.10",
+        "@types/ace-diff": "^2.1.4",
+        "@types/canvas-gauges": "^2.1.8",
+        "@types/flot": "^0.0.36",
+        "@types/flowjs": "^2.13.14",
+        "@types/jquery": "^3.5.31",
+        "@types/js-beautify": "^1.14.3",
+        "@types/leaflet": "1.9.12",
         "@types/leaflet-polylinedecorator": "1.6.4",
         "@types/leaflet-providers": "1.2.4",
         "@types/leaflet.gridlayer.googlemutant": "0.4.9",
         "@types/leaflet.markercluster": "1.5.4",
-        "@types/lodash": "^4.14.192",
-        "@types/marked": "^4.0.8",
+        "@types/lodash": "^4.17.10",
         "@types/node": "~18.15.11",
-        "@types/raphael": "^2.3.2",
-        "@types/react": "17.0.37",
-        "@types/react-dom": "17.0.11",
-        "@types/systemjs": "6.13.1",
-        "@types/tinycolor2": "^1.4.3",
-        "@types/tooltipster": "^0.0.31",
-        "@typescript-eslint/eslint-plugin": "5.57.0",
-        "@typescript-eslint/parser": "5.57.0",
-        "compression-webpack-plugin": "^10.0.0",
-        "directory-tree": "^3.5.1",
-        "eslint": "^8.37.0",
+        "@types/raphael": "^2.3.9",
+        "@types/react": "18.3.10",
+        "@types/react-dom": "18.3.0",
+        "@types/systemjs": "6.15.1",
+        "@types/tinycolor2": "^1.4.6",
+        "@types/tooltipster": "^0.0.35",
+        "@typescript-eslint/eslint-plugin": "^8.12.2",
+        "@typescript-eslint/parser": "^8.12.2",
+        "@typescript-eslint/types": "^8.12.2",
+        "@typescript-eslint/utils": "^8.12.2",
+        "directory-tree": "^3.5.2",
+        "eslint": "~9.13.0",
         "eslint-plugin-import": "latest",
-        "eslint-plugin-jsdoc": "latest",
+        "eslint-plugin-jsdoc": "^50.4.3",
         "eslint-plugin-prefer-arrow": "latest",
-        "jasmine-core": "~3.10.1",
-        "jasmine-spec-reporter": "~7.0.0",
-        "karma": "~6.3.9",
-        "karma-chrome-launcher": "~3.1.0",
-        "karma-coverage-istanbul-reporter": "~3.0.3",
-        "karma-jasmine": "~4.0.1",
-        "karma-jasmine-html-reporter": "^1.7.0",
+        "eslint-plugin-tailwindcss": "^3.17.5",
         "ngrx-store-freeze": "^0.2.4",
-        "patch-package": "^6.5.1",
+        "patch-package": "^8.0.0",
+        "postcss": "^8.4.47",
         "postinstall-prepare": "^2.0.0",
-        "protractor": "~7.0.0",
-        "randexp": "~0.5.3",
-        "raw-loader": "^4.0.2",
-        "source-map-explorer": "^2.5.2",
-        "ts-node": "^10.9.1",
-        "typescript": "~4.9.5",
-        "webpack": "5.77.0"
+        "ts-node": "^10.9.2",
+        "typescript": "~5.5.4"
     },
     "resolutions": {
-        "@types/react": "17.0.37",
-        "ace-builds": "1.4.13",
-        "@date-io/core": "1.3.7"
+        "@types/react": "18.3.10",
+        "rc-virtual-list": "3.5.2",
+        "ace-builds": "1.36.2",
+        "tinymce": "6.8.5",
+        "rollup": "4.22.4",
+        "@babel/core": "7.25.2"
     }
 }
