@@ -14,15 +14,15 @@
 /// limitations under the License.
 ///
 
-import {NgModule} from '@angular/core';
-import {CommonModule} from '@angular/common';
-import {HTTP_INTERCEPTORS, HttpClient, HttpClientModule} from '@angular/common/http';
-import {StoreModule} from '@ngrx/store';
-import {EffectsModule} from '@ngrx/effects';
-import {StoreDevtoolsModule} from '@ngrx/store-devtools';
-import {GlobalHttpInterceptor} from './interceptors/global-http-interceptor';
-import {effects, metaReducers, reducers} from './core.state';
-import {environment as env} from '@env/environment';
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { HTTP_INTERCEPTORS, provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
+import { StoreModule } from '@ngrx/store';
+import { EffectsModule } from '@ngrx/effects';
+import { StoreDevtoolsModule } from '@ngrx/store-devtools';
+import { GlobalHttpInterceptor } from './interceptors/global-http-interceptor';
+import { effects, metaReducers, reducers } from './core.state';
+import { environment as env } from '@env/environment';
 
 import {
   MissingTranslationHandler,
@@ -31,24 +31,19 @@ import {
   TranslateModule,
   TranslateParser
 } from '@ngx-translate/core';
-import {TranslateHttpLoader} from '@ngx-translate/http-loader';
-import {TbMissingTranslationHandler} from './translate/missing-translate-handler';
-import {MatButtonModule} from '@angular/material/button';
-import {MatDialogModule} from '@angular/material/dialog';
-import {MatSnackBarModule} from '@angular/material/snack-bar';
-import {FlexLayoutModule} from '@angular/flex-layout';
-import {TranslateDefaultCompiler} from '@core/translate/translate-default-compiler';
-import {WINDOW_PROVIDERS} from '@core/services/window.service';
-import {TranslateDefaultParser} from '@core/translate/translate-default-parser';
-
-export function HttpLoaderFactory(http: HttpClient) {
-  return new TranslateHttpLoader(http, './assets/locale/locale.constant-', '.json');
-}
+import { TbMissingTranslationHandler } from './translate/missing-translate-handler';
+import { MatButtonModule } from '@angular/material/button';
+import { MatDialogModule } from '@angular/material/dialog';
+import { MatSnackBarModule } from '@angular/material/snack-bar';
+import { FlexLayoutModule } from '@angular/flex-layout';
+import { TranslateDefaultCompiler } from '@core/translate/translate-default-compiler';
+import { WINDOW_PROVIDERS } from '@core/services/window.service';
+import { TranslateDefaultParser } from '@core/translate/translate-default-parser';
+import { TranslateDefaultLoader } from '@core/translate/translate-default-loader';
 
 @NgModule({
   imports: [
     CommonModule,
-    HttpClientModule,
     FlexLayoutModule.withConfig({addFlexToParent: false}),
     MatDialogModule,
     MatButtonModule,
@@ -58,8 +53,7 @@ export function HttpLoaderFactory(http: HttpClient) {
     TranslateModule.forRoot({
       loader: {
         provide: TranslateLoader,
-        useFactory: HttpLoaderFactory,
-        deps: [HttpClient]
+        useClass: TranslateDefaultLoader
       },
       missingTranslationHandler: {
         provide: MissingTranslationHandler,
@@ -89,7 +83,8 @@ export function HttpLoaderFactory(http: HttpClient) {
     env.production
       ? []
       : StoreDevtoolsModule.instrument({
-        name: env.appTitle
+        name: env.appTitle,
+        connectInZone: true
       })
   ],
   providers: [
@@ -98,7 +93,8 @@ export function HttpLoaderFactory(http: HttpClient) {
       useClass: GlobalHttpInterceptor,
       multi: true
     },
-    WINDOW_PROVIDERS
+    WINDOW_PROVIDERS,
+    provideHttpClient(withInterceptorsFromDi()),
   ],
   exports: []
 })
