@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 const CompressionPlugin = require("compression-webpack-plugin");
-const JavaScriptOptimizerPlugin = require("@angular-devkit/build-angular/src/webpack/plugins/javascript-optimizer-plugin").JavaScriptOptimizerPlugin;
+const JavaScriptOptimizerPlugin = require("@angular-devkit/build-angular/src/tools/webpack/plugins/javascript-optimizer-plugin").JavaScriptOptimizerPlugin;
 const webpack = require("webpack");
 const dirTree = require("directory-tree");
 const ngWebpack = require('@ngtools/webpack');
