@@ -14,7 +14,7 @@
 /// limitations under the License.
 ///
 
-declare module '!raw-loader!*' {
-  const contents: string;
-  export = contents;
+declare module '*.raw' {
+  const content: string;
+  export default content;
 }
